just fix this whole mess up

+++ src/main/java/vazkii/quark/experimental/world/BigCaveGenerator.java
@@ -2,10 +2,10 @@
  * This class was created by <Vazkii>. It's distributed as
  * part of the Quark Mod. Get the Source Code in github:
  * https://github.com/Vazkii/Quark
- * 
+ *
  * Quark is Open Source and distributed under the
  * CC-BY-NC-SA 3.0 License: https://creativecommons.org/licenses/by-nc-sa/3.0/deed.en_GB
- * 
+ *
  * File Created @ [10/06/2016, 23:43:17 (GMT)]
  */
 package vazkii.quark.experimental.world;
@@ -21,19 +21,19 @@ public class BigCaveGenerator extends MapGenCaves {
 
 	// We need to use a second RNG since changing the amount of calls in the built-in one will change the way the caves generate
 	Random rand2 = new Random();
-	
+
 	@Override
 	protected void recursiveGenerate(World worldIn, int chunkX, int chunkZ, int p_180701_4_, int p_180701_5_, ChunkPrimer chunkPrimerIn) {
 		int i = rand.nextInt(rand.nextInt(rand.nextInt(15) + 1) + 1);
 		rand2.setSeed(worldIn.getSeed()); // new
-		
+
 		if(rand.nextInt(7) != 0)
 			i = 0;
 
 		for(int j = 0; j < i; ++j) {
-			double d0 = (double)(chunkX * 16 + rand.nextInt(16));
-			double d1 = (double)rand.nextInt(rand.nextInt(120) + 8);
-			double d2 = (double)(chunkZ * 16 + rand.nextInt(16));
+			double d0 = chunkX * 16 + rand.nextInt(16);
+			double d1 = rand.nextInt(rand.nextInt(120) + 8);
+			double d2 = chunkZ * 16 + rand.nextInt(16);
 			int k = 1;
 
 			if(rand.nextInt(4) == 0) {
@@ -48,7 +48,7 @@ public class BigCaveGenerator extends MapGenCaves {
 
 				if(rand.nextInt(10) == 0) // this value can't be changed because that would change the amount of RNG calls
 					f2 *= rand.nextFloat() * rand.nextFloat() * BiggerCaves.bigCaveSizeVariance + BiggerCaves.bigCaveSizeBase;
-				
+
 				if(BiggerCaves.generateHugeCaves && rand2.nextInt(BiggerCaves.hugeCaveChance) == 0 && d1 < BiggerCaves.hugeCaveMaxY)
 					f2 = BiggerCaves.hugeCaveSizeBase + rand2.nextFloat() * BiggerCaves.hugeCaveSizeVariance;
